Move compute_e into each protocol files.
Because this function should be changed to adapt for each specification.
